Support (de)serialization to arrow based on Codec
This MR adds basic (de)serialization for arrow. It was originally
written last summer when I explored a datafusion backend. But I never
got around to creating a MR for them. I confident that we will
eventually merge something like this, so I think we should consider
merging it now so that we do not need to keep updating out of tree code.

The apis might need some work, but as they are just internal I think
that can be improved as we actually start using these.