Simplify expressions using Options
Currently for Options we generate code with a lot of case branches, that
query engines tend not to optimize away. But when the expression is
null intolerant one could optimize away. This adds such logic in the
ExprUnparser.

For extra testing it wrapps all ExprEvaluationSuite tests inside a
MapOption. This uncovered that one of the translation we do in
ExprUnparser did not maintain the null intolerant property.

The main concern here is that we might not have enough test coverage and
then this introduces bugs.